(WIP) Add logging functionality.
THIS IS A WORK IN PROGRESS.  DO NOT SUBMIT.

Adds logging functionality - monitoring for events like message deletions,
user joins and leaves, kicks, bans, etc., and reporting them to a logging
channel.  Also includes a !log command to allow authorized users to add their
own freeform log entries, such as to annotate the log.

In addition, the logger has support for flexible filtering, so that messages
from specified channels or users go instead to another channel instead of the
main logging channel.

Change-Id: Ia2740906ddb16a0bafa3e0e1b48443545d6b526e