Add regex-based message filtering
Allows for automated filtering of messages sent to a Discord server with
varying actions, ranging from only logging, to deleting the message with or
without a warning, to also banning the user.  Filters support automatic expiry
or can be made permanent.

Also adds the start of the CryoHandler code to handle quiets, but so far it does
nothing except keep track of which role is to be used.

Finally, reorganizes some code to improve usability.

Change-Id: Ide88d900b9ed5b083f9a356cda1914c22fb93cff