Add message history dump on cmd, general fixes
This commit adds a privileged command that allows a user's message history to be
dumped, in addition to automatically dumping the history of banned users at the
time of ban set.  The classes related to message dumping have likewise been
broken out of the commands.log into its own package under commands.
MessageCache may also get moved out of the commands.log package some day, but
that day is not today - until then, the message dumping command depends on
logging being enabled.

This commit also makes a minor change to logging; the class that "originated" a
log message is now shown in the footer.

Finally, update copyright years in the file header comments, as appropriate.

Change-Id: If4cdfa06d8739b8fd69229762af98a35a44744ad