Log timestamp as text, chg "permission denied" msg
This commit makes a couple fairly minor changes - first, we now send the
timestamp of log messages as text in the MessageEmbed footer instead of setting
the timestamp of the MessageEmbed itself.  This bypasses issues caused by
Discord doing inconsistent time display things.

Also, change the "permission denied" message to something more appropriate to
the Metroid Community Discord (and also centralize its location to make it
easier to change; still to do is make it configurable, probably in the form of
proper internationalization/localization files).

Change-Id: I1a1298a31b9cc73280f60fdb8bdff954dd916823