Update JDA and Discord API; dependencies cleanup; minor enhancements
Updates the JDA version we depend on, which also updates the bot to be compatible with the most recent version of the Discord API - especially important since the old one will be disabled in a day or two.

Also cleans up some of the transitive dependencies from Beancounter, though the solution (filtering in the Maven shade plugin) means that the bot can no longer be "directly" run as an application in most IDEs; instead, please configure your IDE to run the jarfile produced by the `mvn package` command.

Finally, as a minor enhancements, adds support for showing the time that a user was created in the log message when a user joins a server. There's also probably other fixes here by virtue of updating JDA/the Discord API version.

And one more thing - updates the years in the LICENSE.txt file.

Change-Id: Ib1247c8038ae69cb1a182e5f57740ec85ab773b9