Implement support for react-for-roles messages
Implements support for a server containing one-or-more react-for-roles
messages; i.e., messages where any user can react with one or more
configured emote(s) to have the bot toggle the role on you.

Also supports some "smart" functionality, such as ensuring messages are
synchronized with the database on bot start, unknown reactions are
automatically removed, and the bot automatically treats deletion of a
react-to-role message (or removal of its reactions) as equivalent to
the commands to remove the message or remove the emote-to-role
mapping.

Also fixes a bug where the number of weeks since a user joined was
being calculated incorrectly in the LogListener.

Change-Id: I2df25334dd7b0cd7a30b583a65f24a918f424dfa