(WIP) Implement support for react-for-roles msgs
THIS COMMIT IS A WORK-IN-PROGRESS. DO NOT SUBMIT.

Implements support for a server containing one-or-more react-for-roles
messages; i.e., messages where any user can react with one or more
configured emote(s) to have the bot assign a corresponding role.
Likewise, removing the emote will have the bot remove the corresponding
role.

Also fixes a bug where the number of weeks since a user joined was
being calculated incorrectly in the LogListener.

Change-Id: I2df25334dd7b0cd7a30b583a65f24a918f424dfa