Add Starboard functionality
Adds Starboard functionality, which allows messages that reach a certain
number of reactions with a configured emote to be posted to a designated
starboard channel; sorta like publicly-available democratic pinning.

Supports a global server-wide configuration as well as per-channel
override configurations and excluding channels from the Starboard
system entirely. Each configuration supports specifying which channel
promoted messages go to, what emote is used, and how many reactions with
that emote are needed to promote the message.

Change-Id: I84494f6eb26e36d3332e433480fda0c2741bed59